Finish conversion of Kotlin-based Callable tests & ConfigurationContainer to Spek
